requirements.txt
pyautogen
python-dotenv
litellm